pyproject.toml
[project]
name = "intelligent-personal-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat==0.6.1",
    "autogen-core==0.6.1",
    "autogen-ext==0.6.1",
    "fastapi==0.115.12",
    "google-api-python-client==2.170.0",
    "google-auth-oauthlib==1.2.2",
    "httpx==0.28.1",
    "langchain-core==0.3.65",
    "langchain-google-community==2.0.7",
    "openai==1.86.0",
    "pydantic==2.11.7",
    "python-dotenv>=1.1.1",
    "pytz==2025.2",
    "slack-sdk==3.35.0",
    "tavily-python==0.7.5",
    "tiktoken==0.9.0",
    "tzlocal==5.3.1",
    "uvicorn==0.34.3",
]
